Data.Helpers

Documentation

integralBytes

Arguments

:: (Integral a, Bits a, Show a)

=> a
Value to encode

-> ByteString
Returned ByteString

Encode a value of an Integral type as ByteString.

takeOneConduit :: MonadResource m => Conduit i m i

A Conduit drops every element but the first.

runTCPClientNoWait

Arguments

:: (MonadIO m, MonadBaseControl IO m)

=> ClientSettings m
Settings such as host and port.

-> Application m

-> m ()

Run an Application by connecting to the specified server. Does TCP_NOWAIT

(stolen from network-conduit-0.6.1.1)

isOptionArg :: String -> Bool

Predicate to decide whether a command line argument is an optional argument (starts with a dash).