Data.OAFE

Description

OAFE Types and Functions.

Documentation

type OAFEEvaluation el = HashMap VariableName (Vector el)

The result of all past OAFE evaluations: A map from VariableName to the evaluated values (as Vector).

type OAFEEvaluationRequest el = (VariableName, el)

Request to evaluate one OAFE giving a VariableName and a value.

type OAFEEvaluationResponse el = (VariableName, Vector el)

Response of one OAFE evaluation (the VariableName that has been evaluated with and the values (as Vector)).

type OAFEConfiguration el = HashMap VariableName (Vector (el, el))

The OAFE configuration for all OAFEs.

data OAFEReference

A reference to one OAFE and a specific row in the result vector.

Constructors

OAFERef !VariableName !Int

processOAFEEvaluationRequest :: Field el => OAFEConfiguration el -> OAFEEvaluationRequest el -> OAFEEvaluationResponse el

Process one OAFEEvaluationRequest. This is used by the Token. A implementation of the David & Goliath protocol would replace this by the actual sub-protocol call.