Data.RAE.Evaluation

Description

This module provides various RAE evaluation possibilities.

Public API

evalRAE

Arguments

:: Field el

=> RAE RadicalTuple OAFEReference el
The RAC to evaluate.

-> OAFEEvaluation el
Already evaluated OAFE vals.

-> Either RAEEvaluationFailure el
Either error or result.

Evaluate one RAE.

In real-world use by the functionality David.

data RAEEvaluationFailure

Indicates what went wrong when the evaluation fails.

Constructors

UnknownFailure String String

UnknownVariable VariableName String

IndexOutOfBounds VariableName Int String

Important Special Variable Names

_SPECIAL_VAR_OUT_LEFT_ :: VariableName

The special variable that contains the overall output (left part).

_SPECIAL_VAR_OUT_RIGHT_ :: VariableName

The special variable that contains the overall output (right part).

_SPECIAL_VAR_PRE_OUT_LEFT_ :: VariableName

The special variable that contains the output of the last regular RAE, left part.

_SPECIAL_VAR_PRE_OUT_RIGHT_ :: VariableName

The special variable that contains the output of the last regular RAE, left part.

_SPECIAL_VAR_ADDED_PRE_OUT_ :: VariableName

The special variable that contains the added output of the last reguar RAE.

_SPECIAL_DUAL_VAR_OUT_ :: DualVarName

The special dual variable that contains the overall output.

_SPECIAL_DUAL_VAR_PRE_OUT_ :: DualVarName

The special dual variable that contains the output of the last regular DRAE.

Direct Evaluation of RACs and DRACs (for tests/benchmarks)

runRAC

Arguments

:: forall el . Field el

=> RAC el
The RAC to run.

-> OAFEConfiguration el
The OAFEConfiguration

-> VarMapping el
The values of the input variables.

-> Either String el
Either error message or result.

Directly evaluate a RAC.

This is usually only in use in tests and benchmarks.

runDRAC

Arguments

:: forall el . Field el

=> VarMapping el
Input variable values.

-> DRAC el
The DRAC to run.

-> (Maybe el, VarMapping el)
Result or Nothing

Directly evaluate a DRAC.

This is usually only in use in tests and benchmarks.