Data.RAE.Types

Description

Types for RAEs (Randomized Affine Encodings).

Basic Data Types

data RAE r val el

Randomized Affine Encoding.

Constructors

RAE

Fields

raeMulTerms :: [(r val, r val)]

The multiplicate terms, usually as RadicalTuple of OAFE refs.

raeAddTerms :: [val]

The additive terms, usually OAFE references.

raeConst :: el

A constant value (one tuple component of a DRAV).

data DRAE el

Dual Randomized Affine Encoding.

Constructors

DRAE

Fields

draeKeys :: !(DualKeyPair el)

The DualKeyPair (static dual key&dynamic dual key)

draeMulTerms :: !(DList (MulTermRadicals el))

The multipicative terms as MulTermRadicals

draeAddTerms :: !(DList (DualLinearExpr el))

The additive terms

type RAC el = [RACFragment el]

Randomized Affine Circuit.

type RACFragment el = (VariableName, RAE RadicalTuple OAFEReference el)

Randomized Affine Circuit fragment.

type DRAC el = DList (DRACFragment el)

Dual Randomized Affine Circuit.

type DRACFragment el = (DualVarName, DRAE el)

Dual Randomized Affine Circuit fragment.

data MulTermRadicals el

Multiplicative radicals.

Constructors

MulTermRadicals

Fields

mtrLeft :: DualLinearRadicals el

for left RAC

mtrRight :: DualLinearRadicals el

for right RAC

newtype RadicalTuple a

The radicals travelling separated from each other in tuples.

Constructors

RT (a, a)

newtype RadicalSingleton a

The radicals already unified to one element.

Constructors

RS a

newtype DualLinearRadicals el

Constructors

DLR

Fields

unDLR :: (RadicalTuple (LinearExpr el), RadicalTuple (LinearExpr el))

Encryption Key Data Types

type Key el = el

An encryption key.

type DualKey el = (Key el, Key el)

A dual encryption key (e.g. the dynamic keys or the static keys).

type DualKeyPair el = (DualKey el, DualKey el)

A dual encryption key pair made of two dual keys. Usually the static dual key and the dynamic dual key.

Other Types

data DualVarName

Constructors

DualVarName

Fields

dvnVarName :: !VariableName

dvnLeftVarName :: !VariableName

dvnRightVarName :: !VariableName

Helper Functions

genDualVarName :: VariableName -> DualVarName

Convenience Re-exports

type VariableName = Text

type VarMapping el = Map VariableName el

Variable Mapping

Maps a value to a VariableName

data PrimaryExpression el

Constructors

Variable VariableName

Constant el

Only Exported for Tests

leftVar :: VariableName -> VariableName

rightVar :: VariableName -> VariableName