Functionality.AllInOne

Description

The all in one functionality. This functionality evaluates an Expr just as the usual functionalities would but does eveything in one binary.

This functionality is not meant for productition. It does not have the networking nor the serialization/deserialization overhead.

Documentation

evaluateExpr

Arguments

:: forall el . (Show el, CRandom el, Field el)

=> VarMapping el
David's input variable value(s).

-> Expr el
The Expr Goliath defined.

-> (String -> IO ())
Logging function.

-> IO (Maybe el)
Maybe a result value.

This evaluates an Expr as usual but without network communication in one binary.

This is not used for real-world programs, only for tests, benchmarks and the AllInOne binary.