Functionality.David

Documentation

runRACEvaluation

Arguments

:: forall el . Field el

=> VarMapping el
the initial vars

-> TBMChan (OAFEEvaluationRequest el)
chan to F_OAFE

-> TBMChan (OAFEEvaluationResponse el)
chan from F_OAFE

-> TBMChan (RACFragment el)
RAC streaming

-> TMVar (Maybe el)
final result

-> (String -> IO ())
logging function

-> IO ()

This is the main functionality of David: Evaluating AREs.

It pulls the next ARE from the channel cRACFrag. After having evaluated the ARE, its output gets sent to the Token. The communication threads (see spawnCommThreads) do the real work. This function communication with the communication threads via the channels reqs and rsps. The OAFE evaluation requests get pushed to reqs, after the Token has returned the evaluations, they are received from rsps and saved in a Map.